create git handler and move git related stuff to it
